chore: weather easter egg restructure (#26)
* feat: new custom hook: useWeather
feat: new component: WeatherEasterEgg
chore: moved dark default flares to default-flares

* chore: moved dark default flares back to aurora-components
