+++ src/components/content/heroes/aurora/weather-easter-egg/weather-easter-egg.tsx
@@ -1,0 +1,7 @@
+import { flaresByWeather } from '../aurora-components';
+import { WeatherType } from '../aurora-types';
+import React from 'react';
+
+export const WeatherEasterEgg = ({ weather }) => {
+  return <>{flaresByWeather[weather || WeatherType.NotSet]}</>;
+};
